Rebrand "Cray Shasta" as "HPE Cray EX".
We have an official brandname for Shasta now: HPE Cray EX.  Use that in
the appropriate places, and change the internal name from cray-shasta to
hpe-cray-ex as well.

While here, in the cross-compiling logic in the third-party Makefiles,
use hpe-cray-ex-unknown-linux-gnu instead of ...-cle-linux-gnu for the
configure host type (target type, in Chapel parlance).  This is the same
thing we do for Cray CS targets, and is a better match for EX systems
than what we do for XC ones.