Work-around TOML bug in mason and add a future for bug
TOML can produce a seg fault when parsing a file containing only a comment.
Mason search added a cache file containing only a comment if a cache
file did not exist. To work around this bug, we just create an empty cache file.

Also added a future for the TOML bug