Add the ability to skip printing symbols in internal modules
Add an optional argument to the "get visible symbols" primitive named
`includeInternalModules` that must be assigned a literal `true` or `false`
if it is used. If used, it must also use a named expression.

If it is true, skip printing any found symbols that are in internal modules.
If it is unused or false, print all found symbols.

Add tests with it set either way.