Set: Fix deadlock in isDisjoint()
While holding the parSafe lock in isDisjoint(), don't call size or
contains(), which also take the lock.  Instead, call _size and
_contains(), added here, which rely on the caller having taken the
lock.

Remove the optimization on other.size.  We're about to iterate over
other, so if it's empty we won't be looping over anything.  And the
check introduces the potential for deadlock between a.isDisjoint(b)
and b.isDisjoint(a) if both are parSafe, without a way to establish a
lock ordering between them.

Modify the test/library/standard/Set/isDisjoint/setIsDisjoint test to
run with parSafe=true and with parSafe=false.

closes #16351