Reorder code in atomics support, for symmetry.
The code for ordered and unordered atomics did effectively the same
checks for single-locale execution, types not supported for native
atomics, and atomics against variables not in registered memory, but did
those checks with opposite boolean polarity.  Here, reorder the code so
it looks pretty much the same in both places.  There are no functional
changes here, or at least shouldn't be!

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>