Specify that the comm layer wants the domain threading model.
Libfabric supports several threading models, which specify whether the
provider or the application is responsible for managing contention among
multiple threads for various kinds of fabric resources.  We have not
been specifying a threading level in our hints, with the result that
we've been getting FI_THREAD_SAFE, which says that the provider has all
the responsibility and the application doesn't have to do anything.  But
in fact, the comm layer is written to manage contention for resources
(endpoint, CQ, and so on) itself.  So here, switch to FI_THREAD_DOMAIN
which basically says the opposite, that the application has all of the
responsibility and the provider has none (and thus doesn't have to lock
anything).

Using the verbs;ofi_rxm provider on a Cray CS system, this improves
many-to-one (FastOn) performance by about 15% and empty remote task
spawn time by about 25%.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>