Disable cache remote for tests that check specific optimizations
Throw `--no-cache-remote` for some tests that are using comm counts as a
proxy to see that other optimizations are working. i.e. tests that use
comm counts to check that things like RVF or the RAD-opt are working. In
this case we want to verify the optimizations themselves are firing and
we don't want --cache-remote to hide comms if they aren't.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>