Disable cache remote for distributions comm count tests
These tests are trying to see how much communication different
distribution implementations generate. They are mostly comparing across
implementations or are just trying to track the types of comm patterns
we generate. We don't necessarily care about the exact number of comms
going across the wire, we're just making sure that things look right in
general and we when we change the implementation we haven't added a
drastic amount of communication anywhere.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>