Skip this test on cygwin
With the removal of the single-task optimization in PR #16401 (which
this test relied on heavily), the running time for this test tipped
over the 5-minute mark, presumably due to the higher overhead for
allocating endcounts on Cygwin.  This isn't a particularly important
test to do for cygwin, so skip it.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>