Update error message when conflict due to re-exporting
It's possible that the conflicting symbol that was re-exported had been
re-exported multiple times.  In this case, looking at the re-export point could
be confusing for the user, since they'll be told about it and the original
module where the symbol was defined.  By adjusting the error message to say it
was the last point the symbol was re-exported from before the conflict, the user
is more likely to be open to the possibility that the symbol was re-exported
multiple times.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>