Add parallel extension methods to the `list` type, with tests
Add some methods to `list` designed to make it easier to safely
modify lists in a parallel environment.

The `this()` accessor method is now disabled when a list has been
initialized with `parSafe=true`. The methods `getValue()` and
`getBorrowed()` have been added to make it possible to get a copy of
list elements when `parSafe=true`.

The `set() method has been added to make it possible to set a list
value when `parSafe=true`.

The `computeIndex()` method has been added as a way to access a
reference to a list element in a guarded manner. It is intended for
use with lists that have been initialized with `parSafe=true`. While
it can be used with `parSafe=false` lists, use of `this()` should be
preferred as it is more idiomatic.

Tests have been added for each new method and the `this()` compiler
warning.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>