Adjust test for getValue and add test for error when parSafe this
The method `getValue()` was only able to be called when the map's
value type was a non-nilable class.

This restriction doesnt' make sense (and was pointed out earlier),
so adjust it now. Only emit a compiler error when the type is not
copyable.

Add a compiler warning when attempting to index a map that has been
initialized with `parSafe=true`. Add a test for this warning.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>