Respond to reviewer feedback
Add compiler warnings for `list.first()` and `list.last()` for lists
initialized with `parSafe=true`. Add tests.

Rename `computeIndex()` to `update()`. Add code to `update()` which
relies on the Reflection module to print a prettier error message
when worker `this()` methods fail to resolve. Add a test for this
new error message.

Tweaks to documentation for `update()` and fix a typo in `set()`.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>