Add parallel-safe update() method to the `map` type (#16465)
Add parallel-safe update() method to the `map` type (#16465)

Adjusts the map type by adding a new `update()` method. This method
allows modification of a reference to a map value in a parallel-
safe manner.

The old `update()` method that adds keys and values from another
map has been renamed to `extend()`. This better reflects its
purpose.

Adds compiler warnings to `this()` and `getReference()` that fire
for maps initialized with `parSafe=true`. Add tests for these
new warnings.

Adjust the `getValue()` method to work on all copyable types.
Adjust the `getReference()` method to work on all types.

Adjust the MasonTest module to use maps initialized with
`parSafe=false` for now.

Adjust the UnitTest module to replace several direct indexing calls
with calls to `set()` and `addOrSet()`.

---

Testing:

- [x] ALL on linux64 when COMM=none
- [x] Updated docs rendered locally

Reviewed by @daviditen. Thanks!

---

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>