Fix test for local arrays
I used isDROrDRView to determine whether or not an array was a
single-locale rectangular array, but it turns out that the meaning of
this routine has changed such that for an array 'var B =
A.reindex(...);' or 'var B = A[.., 1]', it no longer returns true for
'B'.  So here, I'm switching to asking whether or not it's rectangular
and whether or not it's a layout instead.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>