Move LinkedLists module to packages
The LinkedLists module has been an old, not-well-maintained outlier
module for quite some time, and has retained its position in
modules/standard primarily due to its historical use within the
internal modules.  But now that we've removed that dependence, it
seems it would be more appropriate in the packages/ directory (not to
say that a linked list implementation couldn't/wouldn't make sense in
the standard modules directory, just that this doesn't seem to be it).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>