Adjust map.items() to return by value rather than const ref
The map.items() iterator currently returns a tuple by const ref. The
current rules for ref tuple returns mean that it is only possible
to return a tuple lvalue by ref, however the items() iterator
returns a tuple expression. This should be a bug (for now), so
adjust items to return by value instead.

Adjust the docs to mention that map.items() returns a copy.

Have items() emit a compiler error if either the keyType or valType
are not copyable.

Remove the overload of items() for non-nilable classes. It is
no longer needed now that Map is implemented using chpl__hashtable.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>