Fix a thinko in previous version: Can still determine ==, != for diff SHAs
My previous draft of this had a bug in that I threw up my hands whenever
two versions differed only by SHA.  However, we can tell that those
things are not equal, so shouldn't be quite so quick to give up.

The approach I took here was to have the spaceship() helper return a
sentinel for that case and then let the outer routines figure out
how to deal with it.  This also helped make the error messages
cleaner again in an `--ignore-errors` setting.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>