Adjust map.items() to return by value rather than const ref (#16494)
Adjust map.items() to return by value rather than const ref (#16494)

The map.items() iterator currently returns a tuple by const ref. The
current rules for ref tuple returns mean that it is only possible
to return a tuple lvalue by ref, however the items() iterator
returns a tuple expression. This should be a bug (for now), so
adjust items() to return by value instead.

Adjust the docs to mention that map.items() returns a copy.

Have items() emit a compiler error if either the keyType or valType
are not copyable.

Remove the overload of items() for non-nilable classes. It is
no longer needed now that Map is implemented using chpl__hashtable.

Resolves #16428.

---

Testing:

- [x] `library/standard/Map` locally with COMM=none
- [x] Updated docs rendered locally

Reviewed by @e-kayrakli. Thanks!

---

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>