Fix block na-none comm counts
While doing a sanity check on master against all my PRs tonight,
I found that PR #16477 improved comm counts for a handful of
tests.  I had the energy to update the block na-none cases, but
not the others, hoping that this would save us at least a few
emails tonight.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>