Force remote atomics to complete before creating tasks.
Add a hook to be called when a child task is created and, in the ofi
comm layer, force remote atomics with delayed completions to be complete
at that time.

While here, move the delayed-completion 'done' indicator directly into
the task-private data instead of allocating it separately and pointing
to it from task-private data.  This saves an allocate and free per task.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>