Update tcl modulefile generation.
Remove references to the PrgEnv- modules, which are already a problem
in some of our target environments such EX with COS v1.3 and later when
using the Lmod module system and, we understand, will become so in
others fairly soon.  To that end, don't load a PrgEnv module for the
user.  Have them do so themselves.  (In all likelihood they have one
loaded already anyway.)  And for EX systems, don't force PrgEnv-gnu any
more, because we're going to start building the module with all the
target compilers.

While here, remove references to the XTPE_NETWORK_TARGET environment
variable and xt-mpich2 module, neither of which have existed for a long
time.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>