Begin integrating all-ref tuples into cullOverReferences
Buuuuuuut, I broke const checking for everything else. Need the diffs
to see what's going on.

These new ref tuples are posing quite a few problems for the
cullOverReferences pass.

The first and most obvious problem is that context calls aren't being
lowered properly. Because the ref tuple itself is considered a
QUAL_VAL, the compiler will prefer USE_VALUE when resolving context
calls.

This doesn't make any sense at all, because a value tuple happens to
have a completely different type than the all-ref version.

So when resolving a context call when the return type is a ref tuple,
the only options that make sense are USE_REF or USE_CONST_REF.

I've forced the compiler to make a selection between those two, but
the problem remains that the selection depends on constness, and
I'm not sure that the constness of ref tuples is being detected
properly.

TLDR: Context calls for ref-tuples can never be USE_VALUE, and I'm
not sure that constness is being propagated correctly because the
compiler views the tuple as a value (it's a blob of references).

Oh, and I broke everything else and need to view the diffs.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>