Revert code in symExprIsSetByDef
Bad nesting of if-statements was causing the cascade of errors.

But now I'm getting different errors related to the constness of
method receivers.

Which, I suspect, is the incorrect constness of ref tuples coming
back to bite me. The compiler is probably preferring USE_CONST_REF
when lowering context calls because it doesn't know any better.

Which in turn makes the array appear const.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>