Double revert changes to doNotChangeTupleTypeRefLevel
Meaning, revert things to the way they were on master. I view this
function as saying "if the called FnSymbol* is a special tuple
intrinsic, don't touch it" _or_ "if this is an iterator, don't touch
it". My hope is that the latter rule can be removed (eventually,
when iterator yield intents are touched up on).

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>