Don't try to default-load the PrgEnv-gnu module on EX systems.
The PrgEnv meta-modules do not exist on EX v1.3 and later.  So instead
of default-loading PrgEnv-gnu on EX, simply insist that the user have
some cpe module loaded.  (The cpe modules are what set PE_ENV in that
environment.)  It's not practical to load one ourselves because the cpe
modules are not meta-modules like the PrgEnv ones, and we'd have to load
all the individual things that go with it.

Note that this change is being done without having been tested, because
I'm getting errors installing the Chapel RPM on a EX UAI due to missing
dependencies and the person I need to help me with that isn't available
right now.  But the output of `git diff -b` seems reassuring.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>