Make variables single-definition, adapting to gcc 10.
There were a couple of places in comm=ofi where we were relying on gcc
-fcommon behavior, which was the default before gcc 10.  But gcc 10,
which we need for EX builds, switched to a default of -fno-common.  To
adapt to that, here make the comm-ofi-internal.h declarations of the few
variables global to the entire ofi comm layer externs, and add
definitions of those in the main comm-ofi.c file.  See #15763 for
earlier, similar changes elsewhere in the runtime.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>