Modernize PALS launcher command line.
The PALS launcher on HPE Cray EX systems no longer needs, or indeed
accepts, the '--' separator between the mpiexec args and the executable
to be launched.  Also, we should throw '--abort-on-failure' to make sure
that when one program instance exits with a non-zero status the entire
job gets shut down right away.  This corresponds to '--kill-on-bad-exit'
which we throw to Slurm srun.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>