Correct codepoint indexing example in String documentation
The example that's described as returning the second codepoint was a
copy of the example returning the second byte.  Change its name and
indexing type to codepoint.

Also, use index 1 to return the second of each.

Also, each type of indexing returns a string, so the functions'
declaration as returning int failed to compile.  I opted to remove the
declared return type instead of specifying string.

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>