Update these tests for the new behavior and remove the .future file
These tests were locking in the warning message when a type only has a secondary
initializer defined in another module, which was a problem during the transition
between constructors and initializers.  Now that initializers are fully
supported and constructors are no longer supported, the warning is no longer
necessary so was removed from the compiler.  As a consequence, these tests no
longer need to expect that warning message.  Additionally, the future
inDiffModule2.chpl had failed to be updated due to not having a .bad file and
not adjusting for owned being the default for classes - so update the test to
no longer delete an owned instance and remove the .future because it now passes.
Also added a call to the generated default initializer, because that should
always be present.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>