Change how the types of nested tuple elements are computed
This is a big change that I've been kind of afraid to make, but the
test `release/examples/primers/reductions.chpl` execution failure
forced my hand.

Previously, the compiler would only recompute the type
`_ref((int, int))` to the type `(_ref(int), _ref(int))` for tuples
labeled `ref`.

Now, the compiler does this unconditionally. This means for _any_
tuple that is not a value tuple, the type of a tuple element is
flattened to be an all-ref tuple.

The amazing thing is how gracefully the compiler handled this change
in types. As soon as I made this change, my test program _and_
the reductions primer just _worked_!

The compiler actually inserted the necessary coercions _inside_ the
generated `_build_tuple` function to adjust tuple ref levels.

This actually makes me think that it might be possible for me to
remove the `chpl_refTupleRepack` function entirely in favor of
coercions.

First things first, time to rerun tests to make sure that everything
is still working.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>