Adjust case for PRIM_SET_MEMBER in symExprIsSetByUse
The code used to return true when the field being set _was_ a ref,
and did _not_ check to make sure call->get(1) was the use.

Change it so that it returns true if call->get(1) is a use and
call->get(2) is not a reference.

Since I'm not sure if this is correct in all cases (only testing will
tell), leave the old code behind in a comment.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>