Default to CHPL_LIBFABRIC=system whenever possible.
We've been defaulting to CHPL_LIBFABRIC=libfabric, to use our bundled
libfabric, everywhere but on HPE Cray EX systems.  That's arguably not
correct.  If there is a system libfabric available we should use it in
preference to the bundled one, because it's more likely to be suitable
to the system at hand.  So here, change the default: if pkg-config is
available and can find a libfabric, use that one.  Only use the bundled
one otherwise.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>