Fix a race in counting running tasks.
In the runtime's pre-user-code hook we force the running task counts to
the proper values after module initialization is done, because (for
reasons we need not go into here) that phase can leave the counts
incorrect but we need them to be correct when we embark on the user
program.  We also set up memory tracking in that hook, because we don't
want to track memory during module init but we do want to track it (if
doing so is requested) in the user code.  But this has a race with the
running task count setting, due to on-stmts from non-0 nodes back to
node 0 to access the associated config const values.  Here, add a
barrier between the running task count reset and the memory tracking
setup to resolve this race.

While here, improve the commentary that describes why the pre-user-code
hook is the way it is.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>