Repair delayed GET visibility.
The delayed PUT visibility commit had some flaws.  Here, fix those.

Clean up the code that decides whether PUTs will be made visible by
means of delivery-complete, immediate following GETs, or delayed
following GETs, and make sure that in this last case we don't return
before the PUT data has been injected.  (This was at the root of an
ft_transposed verifcation failure I've been chasing.)

Force PUTs into visibility before AMOs, except for READ AMOs which
cannot store to the target variable and thus lack MCM implications.
They also have to be forced into visibility before other PUTs which
might overlap, done via AMs.

In addition to PUTs, we also have to ensure that the results of
non-fetching AMOs are visible before we enter a barrier.  But we don't
have to ensure that the PUTs to allow our child nodes in the barrier
tree to proceed are visible -- that can happen as they occur.

While here, improve comments noting that the results of AMs that do RMA
PUT and AMO operations must be visible before their 'done' indicators
are.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>