Simplify and correct message ordering specification.
Our selection of message ordering requirements was both incorrect and
incomplete.  We need write-after-write ordering, or we can't conform to
the same-address clause of the MCM.  But we don't need write-after-read,
because libfabric provides deliver-complete semantics by default for
reads and thus we won't even initiate a following write until after
earlier reads are done.  In addition, there is no need to distinguish
between RMA-only or AMO-only or RMA+AMO read and write ordering based on
whether we're using a provider that can or cannot do native atomics.
The general-case RMA+AMO ordering specifiers cover both cases and are
easier to read and think about in the code.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>