Improve slurm for aws auto-scaling
Stop throwing `--responding` when trying to determine the number of
cores for slurm. With AWS auto-scaling nodes will go offline when they
haven't been used in a while so we want to consider non-responding nodes
too.

This was originally added in d5be4cccd8 but there wasn't any strong
reason for it, there just didn't seem to be a good reason to consider
offline nodes, but there is now.

Without this when using auto-scaling we'd see:
```
warning: Unable to run 'sinfo' (no bytes read)
error: Error trying to determine number of cores per node
```

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>
Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>