Simplify specifying what comm=ofi debug output is desired.
There are lots of different kinds of debug output available from the ofi
comm layer.  Historically you've had to specify a bitmask in the env var
that selects which kinds to produce.  I'm tired of having to look in the
debug support for the bit values, do bitops in my head, and cut+paste to
set this stuff.  Here, change the env var so that it's a comma-separated
list of keywords instead, and arrange for some help output as well.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>