Replace symbolic link python3 with wrapper
The venv normally includes in some-venv/bin/python3 a symbolic link to
the real python3 interpreter used.

This has the drawback that when distributing the venv, the relevant
python executable might be in a different place.

So, this commit adds a bash script that runs python while telling it to
think it is being launched from that location (that is what the exec -a
here does). The wrapper uses whichever python3 is available in the
system. This allows the python3 being run to change (e.g. if the venv is
used on a system with a different path to the python3).

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>