Clean up RPM spec file generation.
We've been producing the RPM spec file by doing shell expansions against
a here-document in the generate-rpmspec.bash script, whose output is the
spec file itself.  Here, simplify and modernize this by using a small
python program to do the necessary textual substitutions, starting with
a template for the spec file.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>