Fix make && make man when CHPL_HOME is not set
Some nightly testing configurations run `make` and then `make man`
without having `CHPL_HOME` set. However, `make man` was recently
changed to use `run-in-venv.bash` but this script required `CHPL_HOME`
be set.

This commit adjusts `run-in-venv.bash` to determine `CHPL_HOME` if it is
not set and in that event to set `CHPL_HOME` for the subcommand. This
allows `make && make man` to function when `CHPL_HOME` is not set.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>