Update version to 1.23.1, preparing for HPE Cray EX v1.3 releases.
EX v1.3 development and maintenance will be based on the 1.23 release
sources rather than the more volatile master branch. Here, prepare for
that by updating the version number.  This makes the same sort of
changes as were done in #16228 which prepared for EX 1.3 in the
release/1.22 branch by bumping to 1.22.2, and as were done in #16578
which recently updated to 1.24.  However, it skips the things the latter
did to put us back into "normal" (pre-release) mode, because for the
purposes of chasing the evolution of EX we're always in release mode.

Signed-off-by: Greg Titus <gbtitus@users.noreply.github.com>