Revert "Remove generated man pages on clean/cleanall/clobber"
This reverts commit 0b4a2c622a76a093174ac3ad414bb5a69b112a14.

It turns out that our module build relies on `make clobber` explicitly leaving
the man pages behind.  We do think it's reasonable for `make clobber` to remove
these files and expect the right thing to do is probably to make a special
clean up command for the module build to call that removes most things except
the built man pages.  But we need to think about that a bit more, so just
revert this change for now.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>

Signed-off-by: CaptainSharf <mohdsharf7407@gail.com>